# Conflicts:
#   test/CMakeLists.txt

+++ src/util/random.h
@@ -4,8 +4,10 @@
 
 namespace ouinet::util::random {
 
-void data(void*, unsigned int);
-std::string string(unsigned int size);
+// NOTE: These are not cryptographically safe.
+
+void data(void*, size_t);
+std::string string(size_t size);
 
 template<typename N /* e.g. uint64_t */>
 inline N number()
